refactor: extract AppState file-reveal actions extension (#590, #430)
Tenth slice of #430 (AppState decomposition). Per #587 design sketch:
extracts 3 file-reveal action delegators (openScreenshot,
showRevealInFinderToast, revealInFinder) into a new
Sources/BugNarrator/AppState+FileRevealActions.swift as
`extension AppState { ... }`.

Range: AppState.swift lines 1122-1142 (21 lines, 729 bytes, SHA-256
19bbecaefe4c3aef32d1ba56ae3200a3964f3220f073b72f652bbe35df7cd539).

**Domain name refined pre-build**: codex spec review flagged the initial
"Toasts" name as weak — 2 of 3 methods (openScreenshot, revealInFinder)
are appUtilityActionPresenter delegators, not toast-related.
Recomputed as "FileRevealActions": all 3 methods cluster around user
file-reveal actions:
- openScreenshot opens a screenshot file
- revealInFinder reveals a file in Finder
- showRevealInFinderToast displays a "Reveal in Finder" toast whose
  action closure delegates back to revealInFinder

That's the coherent domain — primary collaborator (appUtilityActionPresenter
+ appUtilityActions) plus paired auxiliary (transientToastController)
for the toast variant.

All 3 methods pass the delegator test (per the sharpened rule from #589
review 25): single-expression bodies, arguments are observed
(parameters, direct property reads), no branching or derivation.
showRevealInFinderToast has a `[weak self]` capture in a
TransientToastAction trailing closure that calls `self?.revealInFinder(url)`
— cross-file resolution proven by the green build (same type, internal
access preserved, `self` resolution is per-type not per-file).

Full triple-proof stack applied:
- Endpoint verification (1122 = openScreenshot header, 1142 = `}` of
  revealInFinder, 1143 = blank)
- SHA gate matched at both spec-review (against origin) and post-extract
- Byte-diff against origin/main: silent
- Post-deletion double-blank scan: no double-blanks created

Extension inherits @MainActor from AppState per SE-0316 — NOT
re-annotated. New file imports Foundation only (URL).

No dead-import cleanup — AppState.swift still needs Combine +
Foundation.

667 unit tests pass unchanged; docs-sync validation passes.

Pre-build spec reviewed by codex in a fresh session (4 findings folded:
xcodegen target update, domain-name refinement Toasts→FileRevealActions,
[weak self] cross-file resolution note, delegator test rule invocation).
