+++ Sources/BugNarrator/AppState.swift
@@ -1119,28 +1119,6 @@ final class AppState: ObservableObject {
         }
     }
 
-    func openScreenshot(_ screenshot: SessionScreenshot) {
-        appUtilityActionPresenter.present(appUtilityActions.openScreenshot(screenshot))
-    }
-
-    func showRevealInFinderToast(_ message: String, revealing url: URL) {
-        transientToastController.showToast(
-            message,
-            style: .success,
-            durationNanoseconds: 5_000_000_000,
-            action: TransientToastAction(
-                title: "Reveal",
-                accessibilityLabel: "Reveal in Finder"
-            ) { [weak self] in
-                self?.revealInFinder(url)
-            }
-        )
-    }
-
-    func revealInFinder(_ url: URL) {
-        appUtilityActionPresenter.present(appUtilityActions.revealInFinder(url))
-    }
-
     private func prepareErrorPresentationSideEffects() {
         recordingSessionController.stopTimer(resetElapsed: status.phase == .recording)
         recordingSessionController.endActivity()
